Move settings class and helper functions to new module
